Complete deploy v1.1

+++ src/main/java/com/bezkoder/spring/datajpa/controller/CashController.java
@@ -43,14 +43,14 @@ public class CashController {
 
     @PostMapping("/cash")
     public ResponseEntity transferMoney(@RequestBody CashDTO CashDTO) {
-        int userId = CashDTO.getUserId();
-        Optional<User> findUser = userRepository.findById((long) userId);
+        long userId = CashDTO.getUserId();
+        Optional<User> findUser = userRepository.findById( userId);
         if (!findUser.isPresent()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
         User user = findUser.get();
         BigDecimal receiveCash = new BigDecimal(CashDTO.getCash());
-        BigDecimal currentCash = walletRepository.getCurrentCash(userId);
+        BigDecimal currentCash = walletRepository.getCurrentCash((int)userId);
         if (
                 currentCash.compareTo(receiveCash) == -1
                         || receiveCash.compareTo(BigDecimal.ZERO) != 1
